Audit of a dataframe

This audit function is used to assess a data frame. It will return the percentage of values that are missing, where the percentage is dependent on user input. It also returns information on the outlier of the categorcial and nuemric data.

audit(
  data,
  catCut = 0.05,
  levels_numeric = 4,
  levels_categorical = 5,
  row_percent = 35,
  col_percent = 20
)

Arguments

data

a dataframe

catCut

minimum percentage of values in a category in order to not be an outlier

levels_numeric

minimum number of unique values that a numeric variable should have

levels_categorical

maximum number of categories that a categorical variable should have

row_percent

the upper bound percent value for missing values in a row

col_percent

the upper bound percent value for missing values in a column

Value

a message in the console that list an analysis on the missing values of the data, the categorical and numeric variables and the outliers based on these datatypes

Details

Once given a dataframe it will return an audit that describes the data frame.

Examples

audit(mtcars)
#> Descripttion of Dataframe
#> 
#> Number of Observations: 32 
#> 
#> Variable Names: mpg cyl disp hp drat wt qsec vs am gear carb 
#> 
#> Running Missing Values Analysis...
#> 
#> Missing Values Analysis passed with no Warnings
#> 
#> Running Type Check Analysis...
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
#> [1] "The variable cyl has very few unique values. We suggest turing this variable into 3 factor/s."
#> [1] "The variable vs has very few unique values. We suggest turing this variable into 2 factor/s."
#> [1] "The variable am has very few unique values. We suggest turing this variable into 2 factor/s."
#> [1] "The variable gear has very few unique values. We suggest turing this variable into 3 factor/s."
#> 
#> Running Outliers Analysis...
#> Numeric variables that have outliers: 
#> 1 outlier in carb 
#> Categorical variables that have outliers: 
#> Examine these variables and decide whether they should be excluded in further analysis.
audit(airquality)
#> Descripttion of Dataframe
#> 
#> Number of Observations: 153 
#> 
#> Variable Names: Ozone Solar.R Wind Temp Month Day 
#> 
#> Running Missing Values Analysis...
#> There are more than, 20% missing values in the column/s: Ozone.  
#> 
#> Running Type Check Analysis...
#> [1] "Type Check Analysis passed with no Warnings"
#> 
#> Running Outliers Analysis...
#> Numeric variables that have outliers: 
#> 1 outlier in Wind 
#> Categorical variables that have outliers: 
#> Examine these variables and decide whether they should be excluded in further analysis.