airquality Dataset

A dataframe that is derived from information collected about the daily air quality measurements in New York, May to September 1973.

A dataframe that is derived from the 1974 Motor Trend US magazine, and comprises fuel consumption and 10 aspects of automobile design and performance for 32 automobiles (1973–74 models)

mtcars

mtcars

Format

A data frame with 153 rows and 6 columns:

Ozone

numeric. Mean ozone in parts per billion from 1300 to 1500 hours at Roosevelt Island.

Solar.R

numeric. Solar radiation in Langleys in the frequency band 4000–7700 Angstroms from 0800 to 1200 hours at Central Park.

Wind

numeric. Average wind speed in miles per hour at 0700 and 1000 hours at LaGuardia Airport.

Temp

numeric. Charles River dummy variable (= 1 if tract bounds river; 0 otherwise).

Month

numeric. Month(1-12).

Day

numeric. Day of month (1--31).

A data frame with 153 rows and 6 columns:

mpg

numeric. Miles/(US) gallon.

cyl

numeric. Number of cylinders.

disp

numeric. Displacement (cu.in.).

hp

numeric. Gross horsepower.

drat

numeric. Rear axle ratio.

wt

numeric. Weight (1000 lbs).

qsec

numeric. 1/4 mile time.

vs

numeric. Engine (0 = V-shaped, 1 = straight)

am

numeric. Transmission (0 = automatic, 1 = manual)

gear

numeric. Number of forward gears

Source

base R

base R